pd: extract Task
