+++ components/pd/Cargo.toml
@@ -18,6 +18,7 @@ slog = { version = "2.3" }
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "91904ade" }
 tikv_alloc = { path = "../tikv_alloc", default-features = false }
 tikv_util = { path = "../tikv_util" }
+tikv_misc = { path = "../tikv_misc" }
 tokio-timer = "0.2"
 
 [dependencies.prometheus]
