+++ components/pd/src/lib.rs
@@ -36,6 +36,7 @@ pub mod errors;
 pub mod metrics;
 pub mod raft;
 pub mod stats;
+pub mod task;
 pub mod util;
 
 pub type PdFuture<T> = Box<dyn Future<Item = T, Error = Error> + Send>;
