+++ src/pd/pd.rs
@@ -1,6 +1,5 @@
 // Copyright 2016 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::fmt::{self, Display, Formatter};
 use std::sync::Arc;
 
 use futures::Future;
@@ -19,13 +18,11 @@ use pd2::{Error, PdClient, RegionStat};
 use crate::raftstore::coprocessor::{get_region_approximate_keys, get_region_approximate_size};
 use crate::raftstore::store::cmd_resp::new_error;
 use tikv_misc::store_util::is_epoch_stale;
-use tikv_misc::store_util::KeysInfoFormatter;
 use tikv_misc::raftstore_callback::Callback;
 use tikv_misc::store_info::StoreInfo;
 use crate::raftstore::store::{CasualMessage, PeerMsg, RaftCommand, RaftRouter};
 use tikv_misc::flow_stats::FlowStatistics;
 use tikv_util::collections::HashMap;
-use tikv_util::escape;
 use tikv_util::time::time_now_sec;
 use tikv_util::worker::{FutureRunnable as Runnable, FutureScheduler as Scheduler, Stopped};
 use pd2::raft::{
@@ -36,111 +33,8 @@ use pd2::raft::{
     new_merge_request,
 };
 
-/// Uses an asynchronous thread to tell PD something.
-pub enum Task {
-    AskSplit {
-        region: metapb::Region,
-        split_key: Vec<u8>,
-        peer: metapb::Peer,
-        // If true, right Region derives origin region_id.
-        right_derive: bool,
-        callback: Callback,
-    },
-    AskBatchSplit {
-        region: metapb::Region,
-        split_keys: Vec<Vec<u8>>,
-        peer: metapb::Peer,
-        // If true, right Region derives origin region_id.
-        right_derive: bool,
-        callback: Callback,
-    },
-    Heartbeat {
-        region: metapb::Region,
-        peer: metapb::Peer,
-        down_peers: Vec<pdpb::PeerStats>,
-        pending_peers: Vec<metapb::Peer>,
-        written_bytes: u64,
-        written_keys: u64,
-        approximate_size: Option<u64>,
-        approximate_keys: Option<u64>,
-    },
-    StoreHeartbeat {
-        stats: pdpb::StoreStats,
-        store_info: StoreInfo,
-    },
-    ReportBatchSplit {
-        regions: Vec<metapb::Region>,
-    },
-    ValidatePeer {
-        region: metapb::Region,
-        peer: metapb::Peer,
-        merge_source: Option<u64>,
-    },
-    ReadStats {
-        read_stats: HashMap<u64, FlowStatistics>,
-    },
-    DestroyPeer {
-        region_id: u64,
-    },
-}
-
+pub use pd2::task::Task;
 pub use pd2::stats::{StoreStat, PeerStat};
-
-impl Display for Task {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        match *self {
-            Task::AskSplit {
-                ref region,
-                ref split_key,
-                ..
-            } => write!(
-                f,
-                "ask split region {} with key {}",
-                region.get_id(),
-                escape(split_key)
-            ),
-            Task::AskBatchSplit {
-                ref region,
-                ref split_keys,
-                ..
-            } => write!(
-                f,
-                "ask split region {} with {}",
-                region.get_id(),
-                KeysInfoFormatter(&split_keys)
-            ),
-            Task::Heartbeat {
-                ref region,
-                ref peer,
-                ..
-            } => write!(
-                f,
-                "heartbeat for region {:?}, leader {}",
-                region,
-                peer.get_id()
-            ),
-            Task::StoreHeartbeat { ref stats, .. } => {
-                write!(f, "store heartbeat stats: {:?}", stats)
-            }
-            Task::ReportBatchSplit { ref regions } => write!(f, "report split {:?}", regions),
-            Task::ValidatePeer {
-                ref region,
-                ref peer,
-                ref merge_source,
-            } => write!(
-                f,
-                "validate peer {:?} with region {:?}, merge_source {:?}",
-                peer, region, merge_source
-            ),
-            Task::ReadStats { ref read_stats } => {
-                write!(f, "get the read statistics {:?}", read_stats)
-            }
-            Task::DestroyPeer { ref region_id } => {
-                write!(f, "destroy peer of region {}", region_id)
-            }
-        }
-    }
-}
 
 pub struct Runner<T: PdClient> {
     store_id: u64,
